+++ src/main/java/co/lightmasters/haunt/repository/IgnoreRepository.java
@@ -4,5 +4,5 @@ package co.lightmasters.haunt.repository;
 import co.lightmasters.haunt.model.Ignore;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IgnoreRepository extends CrudRepository<Ignore, Integer> {
+public interface IgnoreRepository extends CrudRepository<Ignore, Long> {
 }
