+++ src/main/java/co/lightmasters/haunt/repository/MatchRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MatchRepository extends CrudRepository<Match, Integer> {
+public interface MatchRepository extends CrudRepository<Match, Long> {
     List<Match> findByUsername(String username);
 
     void deleteByUsernameAndSwipedUsername(String username, String swipedUsername);
